Try to detect enums declared within the dtonator file.
Previously if you generated an enum, Foo, with dtonator, and then
later had a property of that type, on the 1st run of dtonator,
it would not detect that type as an enum, because we used only
reflection, and the .class file hasn't been made yet.

On the 2nd run, the .class file would exist, so now the file would
be detected as the right type.

To help make the 1st run more correct, besides using reflection, we
can also looking for the enum type within our other dtos, and see
if it is an enum.