Use an IdentityHashMap to avoid misrecognizing duplicates.
Sometimes a DTO will implement equality, e.g. for dirty checking
on the client-side of the pristine state vs. the current state.

Which is fine, but if two instances of a DTO are passed to the
server-side, even if they are technically equal per equals/hashCode,
we want to treat them as separate entities when they end up in the
database.